There be dragons here: Pragmatic commits ahead
I know that I should be using TDD, especially for Rails development,
however, due to the limited time/resources available to complete this
project, and the simplicity of several of the necessary features to
implement but difficulty to test correctly, I will be going through a
large number of commits just adding functionality. It is my hope that I
will eventually come back and redo these features using proper
techniques. We shall see.